requirements.txt
alpaca-py>=0.8.2
numpy>=1.24
pandas==2.1.1
PyYAML==6.0.1
matplotlib==3.8.0
yfinance>=0.2.54,<2.0
